+++ reporter/index.js
@@ -4,6 +4,7 @@ const fs = require("fs-extra");
 const yaml = require("js-yaml");
 const path = require("path");
 const pug = require("pug");
+const puppeteer = require("puppeteer");
 
 const groupBy = require("lodash/groupBy");
 const flatten = require("lodash/flatten");
@@ -69,6 +70,32 @@ function reporter(args) {
 
     if (log && c.args.html) {
       console.log(html_dump);
+    }
+  };
+
+  c.convertHtmlToPdf = async function (htmlfilename = "inspection.html", pdffilename = "inspection.pdf") {
+    if (c.args.pdf) {
+      const browser = await puppeteer.launch();
+      const pages = await browser.pages();
+      await pages[0].goto("file://" + path.resolve(path.join(c.args.output, htmlfilename)), {waitUntil: 'networkidle0'});
+      await pages[0].pdf({
+        path: path.resolve(path.join(c.args.output, pdffilename)),
+        format: 'A4',
+        displayHeaderFooter: true,
+        headerTemplate: `
+          <div style="width: 100%; font-size: 11px; padding: 5px 5px 0; position: relative;">
+              <div style="bottom: 5px; text-align: center;"><span class="title"></span></div>
+          </div>
+        `,
+        footerTemplate: `
+          <div style="width: 100%; font-size: 11px; padding: 5px 5px 0; position: relative;">
+              <div style="top: 5px; text-align: center;"><span class="pageNumber"></span>/<span class="totalPages"></span></div>
+          </div>
+        `,
+        // this is needed to prevent content from being placed over the footer
+        margin: { top: '1.5cm', bottom: '1cm' },
+      })
+      await browser.close();
     }
   };